Update other-resources.md
Hi Don,
I added a reference to NppLauncher for Notepad Replacement. I developed this tool since some time myself and it does the job for users being afraid to got in the registry. Hope you find it useful.
BTW Plugin is intentionally not in the PluginList because it works only together with a small launcher app that routes the command line and blocks the calling process.
Regards, Mattes